Updated git-tag.sh

+++ scripts/release_support/git-tag.sh
@@ -47,8 +47,11 @@ declare -A COMPONENT_LIST=(
                 else
                         echo "Tag is not successful. Please pass value to GIT_TAG";
                 fi
-
-                        git push origin --delete "$GIT_TAG"
+		if ( params.DEBUG ) {
+		    git push origin --delete "$GIT_TAG"
+		}
+		else 
+			echo "Run in Debug mode if current Git tag needs to be deleted.";
                 popd
 
         done
